refactor(mutex): Replace tokio::sync::Mutex with std::sync::Mutex for git2::Repository
The async version of Mutex is not necessary, and the synchronous version is sufficient.

The advantage of the async version is that it allows us to keep it locked
across an .await point as the tokio documentation says:

> Contrary to popular belief, it is ok and often preferred to use the ordinary
> Mutex from the standard library in asynchronous code.
>
> The feature that the async mutex offers over the blocking mutex is the
> ability to keep it locked across an .await point. This makes the async mutex
> more expensive than the blocking mutex, so the blocking mutex should be
> preferred in the cases where it can be used. The primary use case for the
> async mutex is to provide shared mutable access to IO resources such as a
> database connection. If the value behind the mutex is just data, it’s usually
> appropriate to use a blocking mutex such as the one in the standard library
> or parking_lot.

But we don't need such ability, and it would be better to release the lock as
soon as possible.